Rename from Default to GoPrimitive.
As @paultyng pointed out, the naming of Default made it sound like we
were working with Terraform defaults, not a default type mapping. Rename
the package and the main type to make it clearer we're mapping to Go
primitives.
